refactor pos lib code
