+++ apps/src/lib/node/ledger/shell/governance.rs
@@ -19,8 +19,9 @@ use namada::ledger::protocol;
 use namada::ledger::storage::types::encode;
 use namada::ledger::storage::{DBIter, StorageHasher, DB};
 use namada::ledger::storage_api::{pgf, token, StorageWrite};
+use namada::proof_of_stake::bond_amount;
 use namada::proof_of_stake::parameters::PosParams;
-use namada::proof_of_stake::{bond_amount, read_total_stake};
+use namada::proof_of_stake::storage::read_total_stake;
 use namada::proto::{Code, Data};
 use namada::types::address::Address;
 use namada::types::storage::Epoch;
