+++ apps/src/lib/node/ledger/shell/queries.rs
@@ -68,7 +68,7 @@ where
 mod test_queries {
     use namada::core::ledger::storage::EPOCH_SWITCH_BLOCKS_DELAY;
     use namada::ledger::pos::PosQueries;
-    use namada::proof_of_stake::read_consensus_validator_set_addresses_with_stake;
+    use namada::proof_of_stake::storage::read_consensus_validator_set_addresses_with_stake;
     use namada::proof_of_stake::types::WeightedValidator;
     use namada::tendermint::abci::types::VoteInfo;
     use namada::types::storage::Epoch;
