+++ apps/src/lib/node/ledger/shims/abcipp_shim.rs
@@ -5,7 +5,7 @@ use std::pin::Pin;
 use std::task::{Context, Poll};
 
 use futures::future::FutureExt;
-use namada::proof_of_stake::find_validator_by_raw_hash;
+use namada::proof_of_stake::storage::find_validator_by_raw_hash;
 use namada::proto::Tx;
 use namada::types::hash::Hash;
 use namada::types::key::tm_raw_hash_to_string;
