+++ proof_of_stake/src/queries.rs
@@ -1,0 +1,457 @@
+//! Queriezzz
+
+use std::cmp;
+use std::collections::{BTreeMap, HashMap, HashSet};
+
+use borsh::BorshDeserialize;
+use namada_core::ledger::storage_api::collections::lazy_map::{
+    NestedSubKey, SubKey,
+};
+use namada_core::ledger::storage_api::{self, StorageRead};
+use namada_core::types::address::Address;
+use namada_core::types::dec::Dec;
+use namada_core::types::storage::Epoch;
+use namada_core::types::token;
+
+use crate::slashing::{find_validator_slashes, get_slashed_amount};
+use crate::storage::{bond_handle, read_pos_params, unbond_handle};
+use crate::types::{
+    BondDetails, BondId, BondsAndUnbondsDetail, BondsAndUnbondsDetails, Slash,
+    UnbondDetails,
+};
+use crate::{storage_key, PosParams};
+
+/// Find all validators to which a given bond `owner` (or source) has a
+/// delegation
+pub fn find_delegation_validators<S>(
+    storage: &S,
+    owner: &Address,
+) -> storage_api::Result<HashSet<Address>>
+where
+    S: StorageRead,
+{
+    let bonds_prefix = storage_key::bonds_for_source_prefix(owner);
+    let mut delegations: HashSet<Address> = HashSet::new();
+
+    for iter_result in storage_api::iter_prefix_bytes(storage, &bonds_prefix)? {
+        let (key, _bond_bytes) = iter_result?;
+        let validator_address = storage_key::get_validator_address_from_bond(
+            &key,
+        )
+        .ok_or_else(|| {
+            storage_api::Error::new_const(
+                "Delegation key should contain validator address.",
+            )
+        })?;
+        delegations.insert(validator_address);
+    }
+    Ok(delegations)
+}
+
+/// Find all validators to which a given bond `owner` (or source) has a
+/// delegation with the amount
+pub fn find_delegations<S>(
+    storage: &S,
+    owner: &Address,
+    epoch: &Epoch,
+) -> storage_api::Result<HashMap<Address, token::Amount>>
+where
+    S: StorageRead,
+{
+    let bonds_prefix = storage_key::bonds_for_source_prefix(owner);
+    let params = read_pos_params(storage)?;
+    let mut delegations: HashMap<Address, token::Amount> = HashMap::new();
+
+    for iter_result in storage_api::iter_prefix_bytes(storage, &bonds_prefix)? {
+        let (key, _bond_bytes) = iter_result?;
+        let validator_address = storage_key::get_validator_address_from_bond(
+            &key,
+        )
+        .ok_or_else(|| {
+            storage_api::Error::new_const(
+                "Delegation key should contain validator address.",
+            )
+        })?;
+        let deltas_sum = bond_handle(owner, &validator_address)
+            .get_sum(storage, *epoch, &params)?
+            .unwrap_or_default();
+        delegations.insert(validator_address, deltas_sum);
+    }
+    Ok(delegations)
+}
+
+/// Find if the given source address has any bonds.
+pub fn has_bonds<S>(storage: &S, source: &Address) -> storage_api::Result<bool>
+where
+    S: StorageRead,
+{
+    let max_epoch = Epoch(u64::MAX);
+    let delegations = find_delegations(storage, source, &max_epoch)?;
+    Ok(!delegations
+        .values()
+        .cloned()
+        .sum::<token::Amount>()
+        .is_zero())
+}
+
+/// Find raw bond deltas for the given source and validator address.
+pub fn find_bonds<S>(
+    storage: &S,
+    source: &Address,
+    validator: &Address,
+) -> storage_api::Result<BTreeMap<Epoch, token::Amount>>
+where
+    S: StorageRead,
+{
+    bond_handle(source, validator)
+        .get_data_handler()
+        .iter(storage)?
+        .collect()
+}
+
+/// Find raw unbond deltas for the given source and validator address.
+pub fn find_unbonds<S>(
+    storage: &S,
+    source: &Address,
+    validator: &Address,
+) -> storage_api::Result<BTreeMap<(Epoch, Epoch), token::Amount>>
+where
+    S: StorageRead,
+{
+    unbond_handle(source, validator)
+        .iter(storage)?
+        .map(|next_result| {
+            let (
+                NestedSubKey::Data {
+                    key: start_epoch,
+                    nested_sub_key: SubKey::Data(withdraw_epoch),
+                },
+                amount,
+            ) = next_result?;
+            Ok(((start_epoch, withdraw_epoch), amount))
+        })
+        .collect()
+}
+
+/// Collect the details of all bonds and unbonds that match the source and
+/// validator arguments. If either source or validator is `None`, then grab the
+/// information for all sources or validators, respectively.
+pub fn bonds_and_unbonds<S>(
+    storage: &S,
+    source: Option<Address>,
+    validator: Option<Address>,
+) -> storage_api::Result<BondsAndUnbondsDetails>
+where
+    S: StorageRead,
+{
+    let params = read_pos_params(storage)?;
+
+    match (source.clone(), validator.clone()) {
+        (Some(source), Some(validator)) => {
+            find_bonds_and_unbonds_details(storage, &params, source, validator)
+        }
+        _ => {
+            get_multiple_bonds_and_unbonds(storage, &params, source, validator)
+        }
+    }
+}
+
+fn get_multiple_bonds_and_unbonds<S>(
+    storage: &S,
+    params: &PosParams,
+    source: Option<Address>,
+    validator: Option<Address>,
+) -> storage_api::Result<BondsAndUnbondsDetails>
+where
+    S: StorageRead,
+{
+    debug_assert!(
+        source.is_none() || validator.is_none(),
+        "Use `find_bonds_and_unbonds_details` when full bond ID is known"
+    );
+    let mut slashes_cache = HashMap::<Address, Vec<Slash>>::new();
+    // Applied slashes grouped by validator address
+    let mut applied_slashes = HashMap::<Address, Vec<Slash>>::new();
+
+    // TODO: if validator is `Some`, look-up all its bond owners (including
+    // self-bond, if any) first
+
+    let prefix = match source.as_ref() {
+        Some(source) => storage_key::bonds_for_source_prefix(source),
+        None => storage_key::bonds_prefix(),
+    };
+    // We have to iterate raw bytes, cause the epoched data `last_update` field
+    // gets matched here too
+    let mut raw_bonds = storage_api::iter_prefix_bytes(storage, &prefix)?
+        .filter_map(|result| {
+            if let Ok((key, val_bytes)) = result {
+                if let Some((bond_id, start)) = storage_key::is_bond_key(&key) {
+                    if source.is_some()
+                        && source.as_ref().unwrap() != &bond_id.source
+                    {
+                        return None;
+                    }
+                    if validator.is_some()
+                        && validator.as_ref().unwrap() != &bond_id.validator
+                    {
+                        return None;
+                    }
+                    let change: token::Amount =
+                        BorshDeserialize::try_from_slice(&val_bytes).ok()?;
+                    if change.is_zero() {
+                        return None;
+                    }
+                    return Some((bond_id, start, change));
+                }
+            }
+            None
+        });
+
+    let prefix = match source.as_ref() {
+        Some(source) => storage_key::unbonds_for_source_prefix(source),
+        None => storage_key::unbonds_prefix(),
+    };
+    let mut raw_unbonds = storage_api::iter_prefix_bytes(storage, &prefix)?
+        .filter_map(|result| {
+            if let Ok((key, val_bytes)) = result {
+                if let Some((bond_id, start, withdraw)) =
+                    storage_key::is_unbond_key(&key)
+                {
+                    if source.is_some()
+                        && source.as_ref().unwrap() != &bond_id.source
+                    {
+                        return None;
+                    }
+                    if validator.is_some()
+                        && validator.as_ref().unwrap() != &bond_id.validator
+                    {
+                        return None;
+                    }
+                    match (source.clone(), validator.clone()) {
+                        (None, Some(validator)) => {
+                            if bond_id.validator != validator {
+                                return None;
+                            }
+                        }
+                        (Some(owner), None) => {
+                            if owner != bond_id.source {
+                                return None;
+                            }
+                        }
+                        _ => {}
+                    }
+                    let amount: token::Amount =
+                        BorshDeserialize::try_from_slice(&val_bytes).ok()?;
+                    return Some((bond_id, start, withdraw, amount));
+                }
+            }
+            None
+        });
+
+    let mut bonds_and_unbonds =
+        HashMap::<BondId, (Vec<BondDetails>, Vec<UnbondDetails>)>::new();
+
+    raw_bonds.try_for_each(|(bond_id, start, change)| {
+        if !slashes_cache.contains_key(&bond_id.validator) {
+            let slashes = find_validator_slashes(storage, &bond_id.validator)?;
+            slashes_cache.insert(bond_id.validator.clone(), slashes);
+        }
+        let slashes = slashes_cache
+            .get(&bond_id.validator)
+            .expect("We must have inserted it if it's not cached already");
+        let validator = bond_id.validator.clone();
+        let (bonds, _unbonds) = bonds_and_unbonds.entry(bond_id).or_default();
+        bonds.push(make_bond_details(
+            params,
+            &validator,
+            change,
+            start,
+            slashes,
+            &mut applied_slashes,
+        ));
+        Ok::<_, storage_api::Error>(())
+    })?;
+
+    raw_unbonds.try_for_each(|(bond_id, start, withdraw, amount)| {
+        if !slashes_cache.contains_key(&bond_id.validator) {
+            let slashes = find_validator_slashes(storage, &bond_id.validator)?;
+            slashes_cache.insert(bond_id.validator.clone(), slashes);
+        }
+        let slashes = slashes_cache
+            .get(&bond_id.validator)
+            .expect("We must have inserted it if it's not cached already");
+        let validator = bond_id.validator.clone();
+        let (_bonds, unbonds) = bonds_and_unbonds.entry(bond_id).or_default();
+        unbonds.push(make_unbond_details(
+            params,
+            &validator,
+            amount,
+            (start, withdraw),
+            slashes,
+            &mut applied_slashes,
+        ));
+        Ok::<_, storage_api::Error>(())
+    })?;
+
+    Ok(bonds_and_unbonds
+        .into_iter()
+        .map(|(bond_id, (bonds, unbonds))| {
+            let details = BondsAndUnbondsDetail {
+                bonds,
+                unbonds,
+                slashes: applied_slashes
+                    .get(&bond_id.validator)
+                    .cloned()
+                    .unwrap_or_default(),
+            };
+            (bond_id, details)
+        })
+        .collect())
+}
+
+fn find_bonds_and_unbonds_details<S>(
+    storage: &S,
+    params: &PosParams,
+    source: Address,
+    validator: Address,
+) -> storage_api::Result<BondsAndUnbondsDetails>
+where
+    S: StorageRead,
+{
+    let slashes = find_validator_slashes(storage, &validator)?;
+    let mut applied_slashes = HashMap::<Address, Vec<Slash>>::new();
+
+    let bonds = find_bonds(storage, &source, &validator)?
+        .into_iter()
+        .filter(|(_start, amount)| *amount > token::Amount::zero())
+        .map(|(start, amount)| {
+            make_bond_details(
+                params,
+                &validator,
+                amount,
+                start,
+                &slashes,
+                &mut applied_slashes,
+            )
+        })
+        .collect();
+
+    let unbonds = find_unbonds(storage, &source, &validator)?
+        .into_iter()
+        .map(|(epoch_range, change)| {
+            make_unbond_details(
+                params,
+                &validator,
+                change,
+                epoch_range,
+                &slashes,
+                &mut applied_slashes,
+            )
+        })
+        .collect();
+
+    let details = BondsAndUnbondsDetail {
+        bonds,
+        unbonds,
+        slashes: applied_slashes.get(&validator).cloned().unwrap_or_default(),
+    };
+    let bond_id = BondId { source, validator };
+    Ok(HashMap::from_iter([(bond_id, details)]))
+}
+
+fn make_bond_details(
+    params: &PosParams,
+    validator: &Address,
+    deltas_sum: token::Amount,
+    start: Epoch,
+    slashes: &[Slash],
+    applied_slashes: &mut HashMap<Address, Vec<Slash>>,
+) -> BondDetails {
+    let prev_applied_slashes = applied_slashes
+        .clone()
+        .get(validator)
+        .cloned()
+        .unwrap_or_default();
+
+    let mut slash_rates_by_epoch = BTreeMap::<Epoch, Dec>::new();
+
+    let validator_slashes =
+        applied_slashes.entry(validator.clone()).or_default();
+    for slash in slashes {
+        if slash.epoch >= start {
+            let cur_rate = slash_rates_by_epoch.entry(slash.epoch).or_default();
+            *cur_rate = cmp::min(Dec::one(), *cur_rate + slash.rate);
+
+            if !prev_applied_slashes.iter().any(|s| s == slash) {
+                validator_slashes.push(slash.clone());
+            }
+        }
+    }
+
+    let slashed_amount = if slash_rates_by_epoch.is_empty() {
+        None
+    } else {
+        let amount_after_slashing =
+            get_slashed_amount(params, deltas_sum, &slash_rates_by_epoch)
+                .unwrap();
+        Some(deltas_sum - amount_after_slashing)
+    };
+
+    BondDetails {
+        start,
+        amount: deltas_sum,
+        slashed_amount,
+    }
+}
+
+fn make_unbond_details(
+    params: &PosParams,
+    validator: &Address,
+    amount: token::Amount,
+    (start, withdraw): (Epoch, Epoch),
+    slashes: &[Slash],
+    applied_slashes: &mut HashMap<Address, Vec<Slash>>,
+) -> UnbondDetails {
+    let prev_applied_slashes = applied_slashes
+        .clone()
+        .get(validator)
+        .cloned()
+        .unwrap_or_default();
+    let mut slash_rates_by_epoch = BTreeMap::<Epoch, Dec>::new();
+
+    let validator_slashes =
+        applied_slashes.entry(validator.clone()).or_default();
+    for slash in slashes {
+        if slash.epoch >= start
+            && slash.epoch
+                < withdraw
+                    .checked_sub(
+                        params.unbonding_len
+                            + params.cubic_slashing_window_length,
+                    )
+                    .unwrap_or_default()
+        {
+            let cur_rate = slash_rates_by_epoch.entry(slash.epoch).or_default();
+            *cur_rate = cmp::min(Dec::one(), *cur_rate + slash.rate);
+
+            if !prev_applied_slashes.iter().any(|s| s == slash) {
+                validator_slashes.push(slash.clone());
+            }
+        }
+    }
+
+    let slashed_amount = if slash_rates_by_epoch.is_empty() {
+        None
+    } else {
+        let amount_after_slashing =
+            get_slashed_amount(params, amount, &slash_rates_by_epoch).unwrap();
+        Some(amount - amount_after_slashing)
+    };
+
+    UnbondDetails {
+        start,
+        withdraw,
+        amount,
+        slashed_amount,
+    }
+}
