+++ proof_of_stake/src/tests/state_machine_v2.rs
@@ -32,6 +32,15 @@ use yansi::Paint;
 use super::utils::DbgPrintDiff;
 use crate::parameters::testing::arb_rate;
 use crate::parameters::PosParams;
+use crate::queries::find_delegations;
+use crate::slashing::find_slashes_in_range;
+use crate::storage::{
+    enqueued_slashes_handle, read_all_validator_addresses,
+    read_below_capacity_validator_set_addresses,
+    read_below_capacity_validator_set_addresses_with_stake,
+    read_below_threshold_validator_set_addresses,
+    read_consensus_validator_set_addresses_with_stake,
+};
 use crate::tests::arb_params_and_genesis_validators;
 use crate::tests::utils::pause_for_enter;
 use crate::types::{
@@ -41,10 +50,8 @@ use crate::types::{
 use crate::{
     below_capacity_validator_set_handle, bond_handle,
     consensus_validator_set_handle, delegator_redelegated_bonds_handle,
-    enqueued_slashes_handle, find_slashes_in_range,
-    read_below_threshold_validator_set_addresses, read_pos_params,
-    redelegate_tokens, validator_deltas_handle, validator_slashes_handle,
-    validator_state_handle, RedelegationError,
+    read_pos_params, redelegate_tokens, validator_deltas_handle,
+    validator_slashes_handle, validator_state_handle, RedelegationError,
 };
 
 prop_state_machine! {
@@ -2454,8 +2461,7 @@ impl StateMachineTest for ConcretePosState {
 
                 // Find delegations
                 let delegations_pre =
-                    crate::find_delegations(&state.s, &id.source, &pipeline)
-                        .unwrap();
+                    find_delegations(&state.s, &id.source, &pipeline).unwrap();
 
                 // Apply redelegation
                 let result = redelegate_tokens(
@@ -2607,10 +2613,9 @@ impl StateMachineTest for ConcretePosState {
                     // updated with redelegation. For the source reduced by the
                     // redelegation amount and for the destination increased by
                     // the redelegation amount, less any slashes.
-                    let delegations_post = crate::find_delegations(
-                        &state.s, &id.source, &pipeline,
-                    )
-                    .unwrap();
+                    let delegations_post =
+                        find_delegations(&state.s, &id.source, &pipeline)
+                            .unwrap();
                     let src_delegation_pre = delegations_pre
                         .get(&id.validator)
                         .cloned()
@@ -2725,13 +2730,13 @@ impl ConcretePosState {
         // Post-condition: Consensus validator sets at pipeline offset
         // must be the same as at the epoch before it.
         let consensus_set_before_pipeline =
-            crate::read_consensus_validator_set_addresses_with_stake(
+            read_consensus_validator_set_addresses_with_stake(
                 &self.s,
                 before_pipeline,
             )
             .unwrap();
         let consensus_set_at_pipeline =
-            crate::read_consensus_validator_set_addresses_with_stake(
+            read_consensus_validator_set_addresses_with_stake(
                 &self.s, pipeline,
             )
             .unwrap();
@@ -2743,13 +2748,13 @@ impl ConcretePosState {
         // Post-condition: Below-capacity validator sets at pipeline
         // offset must be the same as at the epoch before it.
         let below_cap_before_pipeline =
-            crate::read_below_capacity_validator_set_addresses_with_stake(
+            read_below_capacity_validator_set_addresses_with_stake(
                 &self.s,
                 before_pipeline,
             )
             .unwrap();
         let below_cap_at_pipeline =
-            crate::read_below_capacity_validator_set_addresses_with_stake(
+            read_below_capacity_validator_set_addresses_with_stake(
                 &self.s, pipeline,
             )
             .unwrap();
@@ -2938,21 +2943,18 @@ impl ConcretePosState {
                 || (num_occurrences == 0 && validator_is_jailed)
         );
 
-        let consensus_set =
-            crate::read_consensus_validator_set_addresses_with_stake(
-                &self.s, pipeline,
-            )
-            .unwrap();
+        let consensus_set = read_consensus_validator_set_addresses_with_stake(
+            &self.s, pipeline,
+        )
+        .unwrap();
         let below_cap_set =
-            crate::read_below_capacity_validator_set_addresses_with_stake(
+            read_below_capacity_validator_set_addresses_with_stake(
                 &self.s, pipeline,
             )
             .unwrap();
         let below_thresh_set =
-            crate::read_below_threshold_validator_set_addresses(
-                &self.s, pipeline,
-            )
-            .unwrap();
+            read_below_threshold_validator_set_addresses(&self.s, pipeline)
+                .unwrap();
         let weighted = WeightedValidator {
             bonded_stake: stake_at_pipeline,
             address: id.validator,
@@ -3015,21 +3017,17 @@ impl ConcretePosState {
                     .contains(address)
             );
             assert!(
-                !crate::read_below_capacity_validator_set_addresses(
-                    &self.s, epoch
-                )
-                .unwrap()
-                .contains(address)
+                !read_below_capacity_validator_set_addresses(&self.s, epoch)
+                    .unwrap()
+                    .contains(address)
             );
             assert!(
-                !crate::read_below_threshold_validator_set_addresses(
-                    &self.s, epoch
-                )
-                .unwrap()
-                .contains(address)
+                !read_below_threshold_validator_set_addresses(&self.s, epoch)
+                    .unwrap()
+                    .contains(address)
             );
             assert!(
-                !crate::read_all_validator_addresses(&self.s, epoch)
+                !read_all_validator_addresses(&self.s, epoch)
                     .unwrap()
                     .contains(address)
             );
@@ -3038,17 +3036,14 @@ impl ConcretePosState {
             crate::read_consensus_validator_set_addresses(&self.s, pipeline)
                 .unwrap()
                 .contains(address);
-        let in_bc = crate::read_below_capacity_validator_set_addresses(
-            &self.s, pipeline,
-        )
-        .unwrap()
-        .contains(address);
+        let in_bc =
+            read_below_capacity_validator_set_addresses(&self.s, pipeline)
+                .unwrap()
+                .contains(address);
         let in_below_thresh =
-            crate::read_below_threshold_validator_set_addresses(
-                &self.s, pipeline,
-            )
-            .unwrap()
-            .contains(address);
+            read_below_threshold_validator_set_addresses(&self.s, pipeline)
+                .unwrap()
+                .contains(address);
 
         assert!(in_below_thresh && !in_consensus && !in_bc);
     }
@@ -3204,7 +3199,7 @@ impl ConcretePosState {
             for WeightedValidator {
                 bonded_stake,
                 address: validator,
-            } in crate::read_consensus_validator_set_addresses_with_stake(
+            } in read_consensus_validator_set_addresses_with_stake(
                 &self.s, epoch,
             )
             .unwrap()
@@ -3274,11 +3269,10 @@ impl ConcretePosState {
             for WeightedValidator {
                 bonded_stake,
                 address: validator,
-            } in
-                crate::read_below_capacity_validator_set_addresses_with_stake(
-                    &self.s, epoch,
-                )
-                .unwrap()
+            } in read_below_capacity_validator_set_addresses_with_stake(
+                &self.s, epoch,
+            )
+            .unwrap()
             {
                 let deltas_stake = validator_deltas_handle(&validator)
                     .get_sum(&self.s, epoch, params)
@@ -3359,10 +3353,8 @@ impl ConcretePosState {
             }
 
             for validator in
-                crate::read_below_threshold_validator_set_addresses(
-                    &self.s, epoch,
-                )
-                .unwrap()
+                read_below_threshold_validator_set_addresses(&self.s, epoch)
+                    .unwrap()
             {
                 let conc_stake = validator_deltas_handle(&validator)
                     .get_sum(&self.s, epoch, params)
@@ -3425,7 +3417,7 @@ impl ConcretePosState {
 
             // Jailed validators not in a set
             let all_validators =
-                crate::read_all_validator_addresses(&self.s, epoch).unwrap();
+                read_all_validator_addresses(&self.s, epoch).unwrap();
 
             for val in all_validators {
                 let state = validator_state_handle(&val)
