+++ tests/src/native_vp/pos.rs
@@ -574,11 +574,11 @@ pub mod testing {
     use namada::proof_of_stake::epoched::DynEpochOffset;
     use namada::proof_of_stake::parameters::testing::arb_rate;
     use namada::proof_of_stake::parameters::PosParams;
-    use namada::proof_of_stake::types::{BondId, ValidatorState};
-    use namada::proof_of_stake::{
+    use namada::proof_of_stake::storage::{
         get_num_consensus_validators, read_pos_params, unbond_handle,
-        ADDRESS as POS_ADDRESS,
     };
+    use namada::proof_of_stake::types::{BondId, ValidatorState};
+    use namada::proof_of_stake::ADDRESS as POS_ADDRESS;
     use namada::types::key::common::PublicKey;
     use namada::types::key::RefTo;
     use namada::types::storage::Epoch;
