+++ tx_prelude/src/proof_of_stake.rs
@@ -5,15 +5,15 @@ use namada_core::types::key::common;
 use namada_core::types::transaction::pos::BecomeValidator;
 use namada_core::types::{key, token};
 pub use namada_proof_of_stake::parameters::PosParams;
-use namada_proof_of_stake::types::ValidatorMetaData;
+use namada_proof_of_stake::storage::read_pos_params;
+use namada_proof_of_stake::types::{ResultSlashing, ValidatorMetaData};
 use namada_proof_of_stake::{
     become_validator, bond_tokens, change_consensus_key,
     change_validator_commission_rate, change_validator_metadata,
     claim_reward_tokens, deactivate_validator, reactivate_validator,
-    read_pos_params, redelegate_tokens, unbond_tokens, unjail_validator,
-    withdraw_tokens,
+    redelegate_tokens, unbond_tokens, unjail_validator, withdraw_tokens,
 };
-pub use namada_proof_of_stake::{parameters, types, ResultSlashing};
+pub use namada_proof_of_stake::{parameters, types};
 
 use super::*;
